navlink performance improvement, remove handleRoute and provide createNavLinkComponent for users to customize it
